+++ sdk/cosmosdb/cosmos/samples/v4/javascript/ContainerManagement.js
@@ -12,6 +12,7 @@ const {
   VectorEmbeddingDataType,
   VectorEmbeddingDistanceFunction,
   VectorIndexType,
+  VectorIndexQuantizationType,
 } = require("@azure/cosmos");
 
 const key = process.env.COSMOS_KEY || "<cosmos key>";
@@ -160,12 +161,14 @@ async function run() {
       {
         path: "/vector2",
         type: VectorIndexType.QuantizedFlat,
+        quantizerType: VectorIndexQuantizationType.Product,
         quantizationByteSize: 2,
         vectorIndexShardKey: ["/Country"],
       },
       {
         path: "/vector3",
         type: VectorIndexType.DiskANN,
+        quantizerType: VectorIndexQuantizationType.Spherical,
         quantizationByteSize: 2,
         indexingSearchListSize: 50,
         vectorIndexShardKey: ["/ZipCode"],