+++ sdk/cosmosdb/cosmos/samples/v4/typescript/src/ContainerManagement.ts
@@ -13,6 +13,7 @@ import {
   VectorEmbeddingDataType,
   VectorEmbeddingDistanceFunction,
   VectorIndexType,
+  VectorIndexQuantizationType,
 } from "@azure/cosmos";
 
 const key = process.env.COSMOS_KEY || "<cosmos key>";
@@ -161,12 +162,14 @@ async function run(): Promise<void> {
       {
         path: "/vector2",
         type: VectorIndexType.QuantizedFlat,
+        quantizerType: VectorIndexQuantizationType.Product,
         quantizationByteSize: 2,
         vectorIndexShardKey: ["/Country"],
       },
       {
         path: "/vector3",
         type: VectorIndexType.DiskANN,
+        quantizerType: VectorIndexQuantizationType.Spherical,
         quantizationByteSize: 2,
         indexingSearchListSize: 50,
         vectorIndexShardKey: ["/ZipCode"],