+++ sdk/cosmosdb/cosmos/src/documents/IndexingPolicy.ts
@@ -71,6 +71,12 @@ export interface VectorIndex {
    */
   type: VectorIndexType;
   /**
+   * The quantizer type for the vector index.
+   * This is an optional parameter and applies to index types DiskANN and quantizedFlat.
+   * Allowed values are "product" and "spherical".
+   */
+  quantizerType?: VectorIndexQuantizationType;
+  /**
    * The number of bytes used in product quantization of the vectors.
    * This is an optional parameter and applies to index types DiskANN and quantizedFlat.
    * The allowed range for this parameter is between 1 and min(Dimensions, 512).
@@ -106,6 +112,21 @@ export enum VectorIndexType {
    * Represents quantizedFlat index type.
    */
   QuantizedFlat = "quantizedFlat",
+}
+
+/**
+ * Represents the quantization type for a vector index.
+ * This is an optional parameter and applies to index types DiskANN and quantizedFlat.
+ */
+export enum VectorIndexQuantizationType {
+  /**
+   * Represents a product quantizer type.
+   */
+  Product = "product",
+  /**
+   * Represents a spherical quantizer type.
+   */
+  Spherical = "spherical",
 }
 
 /**